backend/secure-file-share/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pyotp==2.9.0
cryptography==42.0.2
Pillow==10.2.0
python-magic==0.4.27
django-extensions==3.2.3
pyOpenSSL==25.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1  # Added for runserver_plus